# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Assisted Migration Agent API
  version: v1
servers:
  - url: /api/v1
paths:
...
  /vms/{id}/rightsizing:
    get:
      summary: Get rightsizing utilization breakdown for a specific VM
      operationId: getVMRightsizing
      parameters:
        - name: id
          in: path
          required: true
          description: VirtualMachine MoRef ID
          schema:
            type: string
      responses:
        '200':
          description: VM rightsizing utilization details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmRightsizingDetails'
        '404':
          description: No rightsizing data found for this VM
        '500':
          description: Internal server error
...
  /rightsizing:
    post:
      summary: Trigger rightsizing metrics collection
      operationId: triggerRightsizingCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RightsizingCollectRequest'
      responses:
        '202':
          description: Collection triggered; returns report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingReportSummary'
        '400':
          description: Invalid request
        '500':
          description: Internal server error
    get:
      summary: List all rightsizing reports
      operationId: listRightsizingReports
      responses:
        '200':
          description: List of rightsizing reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingReportListResponse'
        '500':
          description: Internal server error
  /rightsizing/{id}:
    get:
      summary: Get a specific rightsizing report with full VM metrics
      operationId: getRightsizingReport
      parameters:
        - name: id
          in: path
          required: true
          description: Report UUID
          schema:
            type: string
      responses:
        '200':
          description: Rightsizing report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RightsizingReport'
        '404':
          description: Report not found
        '500':
          description: Internal server error
components:
  schemas:
...
    VcenterCredentials:
      required:
        - url
        - username
        - password
      properties:
        url:
          type: string
          format: uri
          description: vCenter URL
          x-oapi-codegen-extra-tags:
            binding: "required,url"
        username:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
        password:
          type: string
          format: password
          minLength: 1
          x-oapi-codegen-extra-tags:
            binding: "required,min=1"
...
    RightsizingCollectRequest:
      type: object
      required:
        - credentials
      properties:
        credentials:
          $ref: '#/components/schemas/VcenterCredentials'
        lookback_hours:
          type: integer
          description: Lookback window in hours (default 720 = 30 days)
          default: 720
        interval_id:
          type: integer
          description: vSphere historical interval ID in seconds (300=day, 1800=week, 7200=month)
          default: 7200
        name_filter:
          type: string
          description: Filter VMs by name substring
        cluster_id:
          type: string
          description: MoRef value of a ClusterComputeResource to scope discovery (e.g. domain-c123)
        batch_size:
          type: integer
          description: Number of VMs per QueryPerf round-trip
          default: 64
        discover_vms:
          type: boolean
          description: >
            If true, discover VMs live from vSphere.
            If false (default), use the VMs already stored in the local inventory.
          default: false
    RightsizingMetricStats:
      type: object
      required:
        - sample_count
        - average
        - p95
        - p99
        - max
        - latest
      properties:
        sample_count:
          type: integer
        average:
          type: number
          format: double
        p95:
          type: number
          format: double
        p99:
          type: number
          format: double
        max:
          type: number
          format: double
        latest:
          type: number
          format: double
    RightsizingVMReport:
      type: object
      required:
        - name
        - moid
        - metrics
        - warnings
      properties:
        name:
          type: string
        moid:
          type: string
        metrics:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RightsizingMetricStats'
        warnings:
          type: array
          description: Non-empty when the VM was queried but had no historical metrics data.
          items:
            type: string
    RightsizingReport:
      type: object
      required:
        - id
        - vcenter
        - cluster_id
        - window_start
        - window_end
        - interval_id
        - expected_sample_count
        - vms
        - created_at
      properties:
        id:
          type: string
          description: UUID of the report
        vcenter:
          type: string
          description: vCenter URL used for this collection
        cluster_id:
          type: string
          description: MoRef of the ClusterComputeResource scoped for this collection (empty = all clusters)
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        interval_id:
          type: integer
        expected_sample_count:
          type: integer
          description: Theoretical maximum samples for the window (lookback / interval)
        vms:
          type: array
          items:
            $ref: '#/components/schemas/RightsizingVMReport'
        created_at:
          type: string
          format: date-time
    RightsizingReportSummary:
      type: object
      description: Rightsizing report metadata without VM metrics. Use GET /rightsizing/{id} to retrieve the full report including metrics.
      required:
        - id
        - vcenter
        - cluster_id
        - window_start
        - window_end
        - interval_id
        - expected_sample_count
        - created_at
      properties:
        id:
          type: string
          description: UUID of the report
        vcenter:
          type: string
          description: vCenter URL used for this collection
        cluster_id:
          type: string
          description: MoRef of the ClusterComputeResource scoped for this collection (empty = all clusters)
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        interval_id:
          type: integer
        expected_sample_count:
          type: integer
          description: Theoretical maximum samples for the window (lookback / interval)
        created_at:
          type: string
          format: date-time
    RightsizingReportListResponse:
      type: object
      required:
        - reports
        - total
      properties:
        reports:
          type: array
          description: List of report summaries. Use GET /rightsizing/{id} for full VM metrics.
          items:
            $ref: '#/components/schemas/RightsizingReportSummary'
        total:
          type: integer
          description: Total number of stored reports
    VmRightsizingDetails:
      type: object
      required:
        - moid
        - vm_name
        - cpu_avg_pct
        - cpu_p95_pct
        - cpu_max_pct
        - cpu_latest_pct
        - mem_avg_pct
        - mem_p95_pct
        - mem_max_pct
        - mem_latest_pct
        - disk_pct
        - confidence_pct
      properties:
        moid:
          type: string
        vm_name:
          type: string
        cpu_avg_pct:
          type: number
          format: double
          description: CPU utilization average (%)
        cpu_p95_pct:
          type: number
          format: double
        cpu_max_pct:
          type: number
          format: double
        cpu_latest_pct:
          type: number
          format: double
        mem_avg_pct:
          type: number
          format: double
          description: Memory utilization average (%)
        mem_p95_pct:
          type: number
          format: double
        mem_max_pct:
          type: number
          format: double
        mem_latest_pct:
          type: number
          format: double
        disk_pct:
          type: number
          format: double
          description: Disk utilization (%)
        confidence_pct:
          type: number
          format: double
          description: Data confidence — sample_count / expected_sample_count × 100